requirements.txt
streamlit>=1.34
pandas>=2.2
numpy
altair>=5.2
Pillow
openpyxl          
pyarrow
minio
tifffile
streamlit
pandas
PyPDF2
streamlit-pdf-viewer